OA_Final/requirements.txt
langchain==0.1.9
langchain-openai==0.0.8
langchain-community==0.0.24
pydantic==2.6.1
streamlit==1.31.1
python-dotenv==1.0.1
mistletoe==1.3.0
openai==1.12.0
typing-extensions==4.9.0
aiohttp==3.9.3
requests==2.31.0